requirements.txt
python-telegram-bot==13.7
requests==2.26.0
telethon
python-magic
libmagic
python-dotenv==1.0.0